+++ src/main/java/com/beehyv/case_study/repositories/MyUserRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MyUserRepo extends JpaRepository<MyUser, Integer> {
-    MyUser getByUserId(int userId);
-    boolean existsByUserId(int userId);
+public interface MyUserRepo extends JpaRepository<MyUser, Long> {
+    MyUser getByUserId(long userId);
+    boolean existsByUserId(long userId);
 }
